Update docs for dwarf line table accelerator peripheral
In the initial PR, the "how to test" section of the docs was a
description of design verification. However this section is intended as
a getting started guide for users. Therefore this change re-writes this
section of the docs to be a user facing example of configuring the
peripheral and running some instructions.